api/model_integration_guide.py
"""
Model Configuration for EmoSense Backend
Instructions for integrating real emotion analysis models
"""

# ======================================================================
# REAL MODEL INTEGRATION GUIDE
# ======================================================================

# 1. TEXT EMOTION MODELS
# =====================

# Option 1: Hugging Face Transformers (Recommended)
# --------------------------------------------------
# Install: pip install transformers torch
# 
# Popular models:
# - "j-hartmann/emotion-english-distilroberta-base" (English emotions)
# - "cardiffnlp/twitter-roberta-base-emotion-multilabel-latest" (Multi-label)
# - "SamLowe/roberta-base-go_emotions" (27 emotions)
# - "michellejieli/emotion_text_classifier" (6 basic emotions)

"""
Example implementation in TextEmotionModel.load_model():

from transformers import AutoTokenizer, AutoModelForSequenceClassification
import torch

self.tokenizer = AutoTokenizer.from_pretrained("j-hartmann/emotion-english-distilroberta-base")
self.model = AutoModelForSequenceClassification.from_pretrained("j-hartmann/emotion-english-distilroberta-base")
self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
self.model.to(self.device)

Example implementation in TextEmotionModel.analyze():

inputs = self.tokenizer(request.text, return_tensors="pt", truncation=True, padding=True, max_length=512)
inputs = {k: v.to(self.device) for k, v in inputs.items()}

with torch.no_grad():
    outputs = self.model(**inputs)
    predictions = torch.nn.functional.softmax(outputs.logits, dim=-1)

emotions = {
    self.model.config.id2label[i]: float(predictions[0][i])
    for i in range(len(predictions[0]))
}
"""

# Option 2: OpenAI API
# -------------------
# Install: pip install openai
# 
"""
import openai

openai.api_key = "your-api-key"

response = openai.ChatCompletion.create(
    model="gpt-3.5-turbo",
    messages=[{
        "role": "user",
        "content": f"Analyze the emotion in this text and return JSON with emotion scores: '{request.text}'"
    }]
)
"""

# Option 3: Google Cloud Natural Language API
# -------------------------------------------
# Install: pip install google-cloud-language
#
"""
from google.cloud import language_v1

client = language_v1.LanguageServiceClient()
document = language_v1.Document(content=request.text, type_=language_v1.Document.Type.PLAIN_TEXT)
response = client.analyze_sentiment(request={"document": document})
"""

# 2. AUDIO EMOTION MODELS
# =======================

# Option 1: Librosa + Custom Models
# ---------------------------------
# Install: pip install librosa scikit-learn
#
"""
import librosa
import numpy as np
from sklearn.ensemble import RandomForestClassifier

def extract_features(audio_data, sr=22050):
    # MFCC features
    mfccs = librosa.feature.mfcc(y=audio_data, sr=sr, n_mfcc=13)
    
    # Spectral features
    spectral_centroids = librosa.feature.spectral_centroid(y=audio_data, sr=sr)
    spectral_rolloff = librosa.feature.spectral_rolloff(y=audio_data, sr=sr)
    
    # Zero crossing rate
    zero_crossing_rate = librosa.feature.zero_crossing_rate(audio_data)
    
    # Combine features
    features = np.concatenate([
        np.mean(mfccs, axis=1),
        np.mean(spectral_centroids),
        np.mean(spectral_rolloff),
        np.mean(zero_crossing_rate)
    ])
    
    return features

# Load audio
audio_data, sr = librosa.load(audio_file, sr=16000)
features = extract_features(audio_data, sr)

# Use pre-trained classifier
emotions = self.classifier.predict_proba([features])[0]
"""

# Option 2: Wav2Vec2 for Speech Emotion Recognition
# ------------------------------------------------
# Install: pip install transformers torch torchaudio
#
"""
from transformers import Wav2Vec2Processor, Wav2Vec2ForSequenceClassification
import torchaudio

processor = Wav2Vec2Processor.from_pretrained("superb/wav2vec2-base-superb-er")
model = Wav2Vec2ForSequenceClassification.from_pretrained("superb/wav2vec2-base-superb-er")

# Process audio
audio_array, sampling_rate = torchaudio.load(audio_file)
if sampling_rate != 16000:
    resampler = torchaudio.transforms.Resample(sampling_rate, 16000)
    audio_array = resampler(audio_array)

inputs = processor(audio_array.squeeze().numpy(), sampling_rate=16000, return_tensors="pt")
logits = model(**inputs).logits
predictions = torch.nn.functional.softmax(logits, dim=-1)
"""

# Option 3: OpenSMILE + Machine Learning
# --------------------------------------
# Install: pip install opensmile
#
"""
import opensmile

smile = opensmile.Smile(
    feature_set=opensmile.FeatureSet.ComParE_2016,
    feature_level=opensmile.FeatureLevel.Functionals,
)

features = smile.process_file(audio_file)
# Use with pre-trained emotion classifier
"""

# 3. VIDEO/IMAGE EMOTION MODELS
# =============================

# Option 1: DeepFace (Recommended for quick start)
# -----------------------------------------------
# Install: pip install deepface
#
"""
from deepface import DeepFace
import cv2

# For image
result = DeepFace.analyze(img_path=image_path, actions=['emotion'])
emotions = result['emotion']

# For video
cap = cv2.VideoCapture(video_path)
while True:
    ret, frame = cap.read()
    if not ret:
        break
    
    try:
        result = DeepFace.analyze(frame, actions=['emotion'], enforce_detection=False)
        emotions = result['emotion']
        # Process frame emotions
    except:
        continue
"""

# Option 2: FER (Facial Expression Recognition)
# --------------------------------------------
# Install: pip install fer
#
"""
from fer import FER
import cv2

detector = FER(mtcnn=True)

# For image
emotions = detector.detect_emotions(image)

# For video
cap = cv2.VideoCapture(video_path)
while True:
    ret, frame = cap.read()
    if not ret:
        break
    
    emotions = detector.detect_emotions(frame)
"""

# Option 3: MediaPipe + Custom Models
# ----------------------------------
# Install: pip install mediapipe opencv-python
#
"""
import mediapipe as mp
import cv2

mp_face_detection = mp.solutions.face_detection
mp_drawing = mp.solutions.drawing_utils

with mp_face_detection.FaceDetection(model_selection=0, min_detection_confidence=0.5) as face_detection:
    image = cv2.imread(image_path)
    results = face_detection.process(cv2.cvtColor(image, cv2.COLOR_BGR2RGB))
    
    if results.detections:
        for detection in results.detections:
            # Extract face region
            # Apply emotion classification model
            pass
"""

# 4. MULTIMODAL MODELS
# ===================

# Option 1: Custom Fusion Architecture
# -----------------------------------
"""
class MultimodalFusion:
    def __init__(self):
        self.text_weight = 0.4
        self.audio_weight = 0.3
        self.video_weight = 0.3
    
    def fuse_emotions(self, text_emotions, audio_emotions, video_emotions):
        # Weighted average
        fused = {}
        all_emotions = set(text_emotions.keys()) | set(audio_emotions.keys()) | set(video_emotions.keys())
        
        for emotion in all_emotions:
            score = (
                text_emotions.get(emotion, 0) * self.text_weight +
                audio_emotions.get(emotion, 0) * self.audio_weight +
                video_emotions.get(emotion, 0) * self.video_weight
            )
            fused[emotion] = score
        
        return fused
"""

# Option 2: Late Fusion with Attention
# -----------------------------------
"""
import torch
import torch.nn as nn

class AttentionFusion(nn.Module):
    def __init__(self, num_emotions):
        super().__init__()
        self.attention = nn.MultiheadAttention(embed_dim=num_emotions, num_heads=1)
        
    def forward(self, text_emb, audio_emb, video_emb):
        # Stack embeddings
        modality_stack = torch.stack([text_emb, audio_emb, video_emb])
        
        # Apply attention
        attended, _ = self.attention(modality_stack, modality_stack, modality_stack)
        
        # Average attended representations
        fused = torch.mean(attended, dim=0)
        
        return torch.softmax(fused, dim=-1)
"""

# 5. DEPLOYMENT CONFIGURATION
# ===========================

# Environment Variables for Model Configuration
MODEL_CONFIG = {
    # Text models
    "TEXT_MODEL_NAME": "j-hartmann/emotion-english-distilroberta-base",
    "TEXT_MODEL_CACHE_DIR": "./models/text",
    
    # Audio models
    "AUDIO_MODEL_NAME": "superb/wav2vec2-base-superb-er",
    "AUDIO_MODEL_CACHE_DIR": "./models/audio",
    "AUDIO_SAMPLE_RATE": 16000,
    
    # Video models
    "VIDEO_MODEL_BACKEND": "deepface",  # or "fer", "mediapipe"
    "VIDEO_MODEL_CACHE_DIR": "./models/video",
    "FACE_DETECTION_CONFIDENCE": 0.5,
    
    # Multimodal fusion
    "FUSION_STRATEGY": "weighted_average",  # or "attention", "voting"
    "TEXT_WEIGHT": 0.4,
    "AUDIO_WEIGHT": 0.3,
    "VIDEO_WEIGHT": 0.3,
    
    # Performance
    "BATCH_SIZE": 32,
    "MAX_SEQUENCE_LENGTH": 512,
    "DEVICE": "auto",  # "cpu", "cuda", or "auto"
    
    # Caching
    "ENABLE_MODEL_CACHING": True,
    "CACHE_TTL": 3600,  # 1 hour
}

# 6. PRODUCTION DEPLOYMENT NOTES
# ==============================

"""
For Production Deployment:

1. Model Storage:
   - Store models in cloud storage (AWS S3, Google Cloud Storage)
   - Use model versioning and A/B testing
   - Implement model warm-up strategies

2. Performance Optimization:
   - Use TensorRT or ONNX for inference optimization
   - Implement model quantization for faster inference
   - Use GPU acceleration when available

3. Scaling:
   - Implement model servers (TorchServe, TensorFlow Serving)
   - Use container orchestration (Kubernetes)
   - Implement load balancing for model inference

4. Monitoring:
   - Track model performance metrics
   - Monitor inference latency and throughput
   - Implement model drift detection

5. Security:
   - Secure model files and API keys
   - Implement input validation and sanitization
   - Use authentication for model endpoints
"""

# 7. QUICK START TEMPLATE
# =======================

"""
To quickly integrate a real model:

1. Choose your model type (text/audio/video)
2. Install required dependencies
3. Update the corresponding model class in emotion_models.py
4. Replace the demo implementation in load_model() and analyze() methods
5. Test with the provided API endpoints
6. Deploy to your preferred platform

Example for text model integration:

# In TextEmotionModel.load_model():
from transformers import AutoTokenizer, AutoModelForSequenceClassification
self.tokenizer = AutoTokenizer.from_pretrained("j-hartmann/emotion-english-distilroberta-base")
self.model = AutoModelForSequenceClassification.from_pretrained("j-hartmann/emotion-english-distilroberta-base")

# In TextEmotionModel.analyze():
inputs = self.tokenizer(request.text, return_tensors="pt", truncation=True, padding=True)
outputs = self.model(**inputs)
predictions = torch.nn.functional.softmax(outputs.logits, dim=-1)
# Convert to emotion dictionary and return EmotionResult
"""
