api/requirements-real-models.txt
# Production requirements with real emotion analysis models
# For deployment environments that support heavier ML dependencies

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Authentication
python-jose==3.3.0
passlib==1.7.4

# Configuration
pydantic==2.4.2
pydantic-settings==2.0.3

# Advanced ML libraries for emotion analysis
torch==2.1.2
transformers==4.36.2
datasets==2.15.0
tokenizers==0.15.0

# Computer Vision for video emotion analysis
opencv-python==4.8.1.78
mediapipe==0.10.8
face-recognition==1.3.0
dlib==19.24.2

# Audio processing for emotion recognition
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.11
webrtcvad==2.0.10
python-speech-features==0.6

# Scientific computing
numpy==1.24.4
scipy==1.11.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0

# NLP libraries
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Image processing
pillow==10.1.0
imageio==2.33.1

# Utilities
tqdm==4.66.1
requests==2.31.0
aiofiles==23.2.1

# Model downloading and caching
huggingface-hub==0.19.4
accelerate==0.25.0

# For multimodal analysis
sentence-transformers==2.2.2
clip-by-openai==1.0

# Performance optimization
onnx==1.15.0
onnxruntime==1.16.3
